+++ src/main/java/com/btl/server/entity/TaiKhoan.java
@@ -14,8 +14,7 @@ public class TaiKhoan implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
+    private Long id; 
     @Column(unique = true, nullable = false)
     private String username;
 
@@ -48,8 +47,8 @@ public class TaiKhoan implements UserDetails {
     @Override
     public boolean isEnabled() { return true; }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
