+++ src/main/java/com/btl/server/repository/TaiKhoanRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.btl.server.entity.TaiKhoan;
 
 @Repository
-public interface TaiKhoanRepository extends JpaRepository<TaiKhoan, Integer> {
+public interface TaiKhoanRepository extends JpaRepository<TaiKhoan, Long> {
      Optional<TaiKhoan> findByUsername(String username);
 }
